projects/dataflow-framework/abstraction-level-6/pyproject.toml
[project]
name = "abstraction-level-6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "typer>=0.17.3",
]
